Moved LegislationScreenData to Data folder
